+++ docs/site_libs/quarto-html/quarto-syntax-highlighting-dark-e2182856f2c32102dd26b05a5ef16cb5.css
@@ -36,94 +36,8 @@
   --quarto-font-monospace: SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
 }
 
-code span.al {
-  font-weight: bold;
-  color: #ff5555;
-}
-
-code span.an {
-  color: #6a737d;
-}
-
-code span.at {
-  color: #f97583;
-}
-
-code span.bn {
-  color: #79b8ff;
-}
-
-code span.bu {
-  color: #f97583;
-}
-
-code span.ch {
-  color: #9ecbff;
-}
-
-code span.co {
-  color: #6a737d;
-}
-
-code span.cv {
-  color: #6a737d;
-}
-
-code span.cn {
-  color: #79b8ff;
-}
-
-code span.cf {
-  color: #f97583;
-}
-
-code span.dt {
-  color: #f97583;
-}
-
-code span.dv {
-  color: #79b8ff;
-}
-
-code span.do {
-  color: #6a737d;
-}
-
-code span.er {
-  color: #ff5555;
-  text-decoration: underline;
-}
-
-code span.ex {
-  font-weight: bold;
-  color: #f97583;
-}
-
-code span.fl {
-  color: #79b8ff;
-}
-
-code span.fu {
-  color: #b392f0;
-}
-
-code span.im {
-  color: #9ecbff;
-}
-
-code span.in {
-  color: #6a737d;
-}
-
-code span.kw {
-  color: #f97583;
-}
-
+/* syntax highlight based on Pandoc's rules */
 pre > code.sourceCode > span {
-  color: #e1e4e8;
-}
-
-code span {
   color: #e1e4e8;
 }
 
@@ -136,38 +50,155 @@ div.sourceCode pre.sourceCode {
   color: #e1e4e8;
 }
 
+/* Normal */
+code span {
+  color: #e1e4e8;
+}
+
+/* Alert */
+code span.al {
+  font-weight: bold;
+  color: #ff5555;
+}
+
+/* Annotation */
+code span.an {
+  color: #6a737d;
+}
+
+/* Attribute */
+code span.at {
+  color: #f97583;
+}
+
+/* BaseN */
+code span.bn {
+  color: #79b8ff;
+}
+
+/* BuiltIn */
+code span.bu {
+  color: #f97583;
+}
+
+/* ControlFlow */
+code span.cf {
+  color: #f97583;
+}
+
+/* Char */
+code span.ch {
+  color: #9ecbff;
+}
+
+/* Constant */
+code span.cn {
+  color: #79b8ff;
+}
+
+/* Comment */
+code span.co {
+  color: #6a737d;
+}
+
+/* CommentVar */
+code span.cv {
+  color: #6a737d;
+}
+
+/* Documentation */
+code span.do {
+  color: #6a737d;
+}
+
+/* DataType */
+code span.dt {
+  color: #f97583;
+}
+
+/* DecVal */
+code span.dv {
+  color: #79b8ff;
+}
+
+/* Error */
+code span.er {
+  color: #ff5555;
+  text-decoration: underline;
+}
+
+/* Extension */
+code span.ex {
+  font-weight: bold;
+  color: #f97583;
+}
+
+/* Float */
+code span.fl {
+  color: #79b8ff;
+}
+
+/* Function */
+code span.fu {
+  color: #b392f0;
+}
+
+/* Import */
+code span.im {
+  color: #9ecbff;
+}
+
+/* Information */
+code span.in {
+  color: #6a737d;
+}
+
+/* Keyword */
+code span.kw {
+  color: #f97583;
+}
+
+/* Operator */
 code span.op {
   color: #e1e4e8;
 }
 
+/* Preprocessor */
 code span.pp {
   color: #f97583;
 }
 
+/* RegionMarker */
 code span.re {
   color: #6a737d;
 }
 
+/* SpecialChar */
 code span.sc {
   color: #79b8ff;
 }
 
+/* SpecialString */
 code span.ss {
   color: #9ecbff;
 }
 
+/* String */
 code span.st {
   color: #9ecbff;
 }
 
+/* Variable */
 code span.va {
   color: #ffab70;
 }
 
+/* VerbatimString */
 code span.vs {
   color: #9ecbff;
 }
 
+/* Warning */
 code span.wa {
   color: #ff5555;
 }
@@ -176,4 +207,4 @@ code span.wa {
   content: "</";
 }
 
-/*# sourceMappingURL=8de98c2b9c397292a29b2bce9eb39a87.css.map */
+/*# sourceMappingURL=0a135725590190ff7f6c7403eceae810.css.map */
